# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "persisted_revision": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "MovePeer": {
        "properties": {
          "group_id": {
            "type": "string"
          }
        },
        "required": [
          "group_id"
        ],
        "type": "object"
      },
...
      "NewPeer": {
        "properties": {
          "group_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "group_id"
        ],
        "type": "object"
      },
      "Peer": {
        "properties": {
          "group_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ipv4": {
            "type": "string"
          },
          "last_handshake_unix": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          },
          "received_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "sent_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "public_key",
          "ipv4",
          "group_id",
          "received_bytes",
          "sent_bytes"
        ],
        "type": "object"
      },
      "PeerProvision": {
        "properties": {
          "config": {
            "type": "string"
          },
          "peer": {
            "$ref": "#/components/schemas/Peer"
          }
        },
        "required": [
          "peer",
          "config"
        ],
        "type": "object"
      },
      "PeerStatus": {
        "properties": {
          "group_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ipv4": {
            "type": "string"
          },
          "last_data_unix": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "last_handshake_unix": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          },
          "received_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "sent_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "public_key",
          "ipv4",
          "group_id",
          "received_bytes",
          "sent_bytes"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminBearer": {
        "description": "Exact configured WIREHUB_ADMIN_TOKEN. Keep it outside URLs and logs.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "wirehub",
    "version": "1.0.0-rc.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/peers": {
      "get": {
        "operationId": "list_peers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PeerStatus"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      },
      "post": {
        "operationId": "create_peer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPeer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeerProvision"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON content type required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid JSON body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      }
    },
    "/api/peers/{id}": {
      "delete": {
        "operationId": "delete_peer",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource missing"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      }
    },
    "/api/peers/{id}/group": {
      "put": {
        "operationId": "move_peer",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovePeer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Peer"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource missing"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON content type required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid JSON body"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Persistence failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Runtime unavailable or activation unconfirmed"
          }
        },
        "security": [
          {
            "adminBearer": []
          }
        ],
        "tags": [
          "crate"
        ]
      }
    },
...
  }
}